% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsureIntercept.R
\name{EnsureIntercept}
\alias{EnsureIntercept}
\title{Ensure constant term in matrix}
\usage{
EnsureIntercept(x)
}
\arguments{
\item{x}{Input matrix}
}
\value{
The input matrix possibly with a column of ones added
}
\description{
A column of ones may be added
}
\examples{
x <- matrix(c(5, 8, 4, 2, 7, 6), 3, 2)
EnsureIntercept(x)
EnsureIntercept(cbind(x, 2))
EnsureIntercept(cbind(x, 0))
EnsureIntercept(matrix(0, 4, 0))
}
\author{
Øyvind Langsrud
}
\keyword{internal}
