% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSDCdata.R
\name{RegSDCdata}
\alias{RegSDCdata}
\title{Function that returns a dataset}
\usage{
RegSDCdata(dataset)
}
\arguments{
\item{dataset}{Name of data set within the RegSDC package}
}
\value{
data frame
}
\description{
Function that returns a dataset
}
\details{
\strong{sec7data:} Data in section 7 of the paper as a data frame

\strong{sec7y:} Y in section 7 of the paper as a matrix

\strong{sec7x:} X in section 7 of the paper as a matrix

\strong{sec7z:} Z in section 7 of the paper as a matrix

\strong{sec7xAll:} Xall in section 7 of the paper as a matrix

\strong{sec7zAll:} Zall in section 7 of the paper as a matrix
}
\examples{
RegSDCdata("sec7data")
RegSDCdata("sec7y")
RegSDCdata("sec7x")
RegSDCdata("sec7z")
RegSDCdata("sec7xAll")
RegSDCdata("sec7zAll")
}
\author{
Øyvind Langsrud
}
