% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSDChybrid.R
\encoding{UTF8}
\name{RegSDChybrid}
\alias{RegSDChybrid}
\title{Regression-based SDC Tools - Generalized microaggregation}
\usage{
RegSDChybrid(y, clusters = NULL, xLocal = NULL, xGlobal = NULL,
  clusterPieces = NULL, xClusterPieces = NULL,
  groupedClusters = NULL, xGroupedClusters = NULL,
  alternative = NULL, alpha = NULL, ySim = NULL,
  returnParts = FALSE, epsAlpha = 1e-07, makeunique = TRUE)
}
\arguments{
\item{y}{Matrix of confidential variables}

\item{clusters}{Vector of cluster coding}

\item{xLocal}{Matrix of x-variables to be crossed with clusters}

\item{xGlobal}{Matrix of x-variables NOT to be crossed with clusters}

\item{clusterPieces}{Vector of coding of cluster pieces}

\item{xClusterPieces}{Matrix of x-variables to be crossed with cluster pieces}

\item{groupedClusters}{Vector of coding of grouped clusters}

\item{xGroupedClusters}{Matrix of x-variables to be crossed with grouped clusters}

\item{alternative}{One of "" (default), "a", "b" or "c"}

\item{alpha}{Possible to specify parameter used internally by alternative "c"}

\item{ySim}{Possible to specify the internally simulated data manually}

\item{returnParts}{Alternative output six matrices:
y1 and y2 (fitted),    e3s and e4s (new residuals),   e3 and e4 (original residuals)}

\item{epsAlpha}{Precision constant for alpha calculation}

\item{makeunique}{Parameter to be used in GenQR}
}
\value{
Generated version of y
}
\description{
Implementation of the methodology in section 6 in the paper
}
\examples{
y <- matrix(rnorm(30) + 1:30, 10, 3)
x <- matrix(1:10, 10, 1)

# Same as RegSDCipso(y)
yOut <- RegSDChybrid(y)

# With a single cluster both are same as RegSDCipso(y, x)
yOut <- RegSDChybrid(y, xLocal = x)
yOut <- RegSDChybrid(y, xGlobal = x)

# Define two clusters
clust <- rep(1:2, each = 5)

# MHa and MHb in paper
yMHa <- RegSDChybrid(y, clusters = clust, xLocal = x)
yMHb <- RegSDChybrid(y, clusterPieces = clust, xLocal = x)

# An extended variant of MHb as mentioned in paper paragraph below definition of MHa/MHb
yMHbExt <- RegSDChybrid(y, clusterPieces = clust, xClusterPieces = x)

# Identical means within clusters
aggregate(y, list(clust = clust), mean)
aggregate(yMHa, list(clust = clust), mean)
aggregate(yMHb, list(clust = clust), mean)
aggregate(yMHbExt, list(clust = clust), mean)

# Identical global regression results
summary(lm(y[, 1] ~ x))
summary(lm(yMHa[, 1] ~ x))
summary(lm(yMHb[, 1] ~ x))
summary(lm(yMHbExt[, 1] ~ x))

# MHa: Identical local regression results
summary(lm(y[, 1] ~ x, subset = clust == 1))
summary(lm(yMHa[, 1] ~ x, subset = clust == 1))

# MHb: Different results
summary(lm(yMHb[, 1] ~ x, subset = clust == 1))

# MHbExt: Same estimates and different std. errors
summary(lm(yMHbExt[, 1] ~ x, subset = clust == 1))

# More examples will be added
}
