% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSDCgen.R
\encoding{UTF8}
\name{RegSDCipso}
\alias{RegSDCipso}
\title{Regression-based SDC Tools - Ordinary synthetic data (IPSO)}
\usage{
RegSDCipso(y, x = matrix(1, NROW(y), 1))
}
\arguments{
\item{y}{Matrix of confidential variables}

\item{x}{Matrix of non-confidential variables (including a constant term (column of ones))}
}
\value{
Generated version of y
}
\description{
Implementation of equation 4 in the paper.
}
\examples{
x1 <- 1:5
x <- cbind(x0 = 1, x1 = x1)
y <- matrix(rnorm(15) + 1:15, 5, 3)
ySynth <- RegSDCipso(y, x)

# Identical regression results
summary(lm(y[, 1] ~ x1))
summary(lm(ySynth[, 1] ~ x1))

# Identical covariance matrices
cov(y) - cov(ySynth)
cov(residuals(lm(y ~ x1))) - cov(residuals(lm(ySynth ~ x1)))
}
