% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateC.R
\name{FindAlpha}
\alias{FindAlpha}
\alias{FindAlphaSimple}
\title{Calculation of alpha}
\usage{
FindAlpha(a, b, tryViaQR = TRUE)

FindAlphaSimple(a, b)
}
\arguments{
\item{a}{matrix E in paper}

\item{b}{matrix Eg in paper}

\item{tryViaQR}{When TRUE QR transformation used (to handle collinearity) 
when ordinary calculations fail.}
}
\value{
alpha
}
\description{
Function to find the largest alpha that makes equation 10 in the paper solvable.
}
\note{
FindAlphaSimple performs the calculations by a simple/direct method. 
FindAlpha is made to handle problematic special cases.
}
\seealso{
See examples in the documentation of \code{\link{CalculateC}}
}
