% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSDCgen.R
\encoding{UTF8}
\name{RegSDCgen}
\alias{RegSDCgen}
\title{Regression-based SDC Tools - General data generation}
\usage{
RegSDCgen(y, x = matrix(1, NROW(y), 1), doSVD = FALSE, yNew = NULL,
  lambda = Inf, makeunique = TRUE)
}
\arguments{
\item{y}{Matrix of confidential variables}

\item{x}{Matrix of non-confidential variables (including a constant term (column of ones))}

\item{doSVD}{SVD when TRUE and QR when FALSE}

\item{yNew}{Matrix of y-data for new scores (simulated when NULL)}

\item{lambda}{ROMM parameter}

\item{makeunique}{Parameter to be used in GenQR}
}
\value{
Generated version of y
}
\description{
IPSO by QR or SVD, scores from arbitrary data, and ROMM
}
\details{
doSVD has effect on decomposition of y and yNew
}
\examples{
exY <- matrix(rnorm(15), 5, 3)
RegSDCgen(exY)
RegSDCgen(exY, yNew = exY + 0.001 * matrix(rnorm(15), 5, 3))  # Close to exY
RegSDCgen(exY, lambda = 0.001)  # Close to exY
}
\keyword{internal}
