% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSDCgen.R
\encoding{UTF8}
\name{RegSDCromm}
\alias{RegSDCromm}
\title{Regression-based SDC Tools - Random orthogonal matrix masking (ROMM)}
\usage{
RegSDCromm(y, lambda = Inf, x = matrix(1, NROW(y), 1), doSVD = FALSE)
}
\arguments{
\item{y}{Matrix of confidential variables}

\item{lambda}{ROMM parameter}

\item{x}{Matrix of non-confidential variables (including a constant term (column of ones))}

\item{doSVD}{SVD when TRUE and QR when FALSE}
}
\value{
Generated version of y
}
\description{
Implementation based on equations 11, 12 and 17 in the paper.
}
\details{
doSVD has effect on decomposition of y and yNew. 
The exact behaviour of the method depends on the choice of the decomposition method because of 
the sequentially phenomenon mentioned in the paper. 
The similarity to the original data will tend to be highest for the first component.
}
\examples{
x1 <- 1:5
x <- cbind(x0 = 1, x1 = x1)
y <- matrix(rnorm(15) + 1:15, 5, 3)

# Same as IPSO (RegSDCipso)
RegSDCromm(y, Inf, x)

# Close to IPSO
RegSDCromm(y, 100, x)

# Close to y
RegSDCromm(y, 0.001, x)
}
