% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressDec.R
\name{SuppressDec}
\alias{SuppressDec}
\title{Title}
\usage{
SuppressDec(x, z, y = NULL, suppressed = NULL, digits = 9,
  nRep = 1, yDeduct = NULL, resScale = NULL, rmse = NULL)
}
\arguments{
\item{x}{}

\item{z}{}

\item{y}{}

\item{suppressed}{}

\item{digits}{}

\item{nRep}{}

\item{yDeduct}{}

\item{resScale}{}

\item{rmse}{}
}
\description{
Title
}
\examples{
# Same data as in the paper
z <- RegSDCdata("sec7z")
x <- RegSDCdata("sec7x")
y <- RegSDCdata("sec7y")  # Now z is t(x) \%*\% y 
zAll <- RegSDCdata("sec7zAll")
zAllSupp <- RegSDCdata("sec7zAllSupp")
xAll <- RegSDCdata("sec7xAll")

# When no suppression, output is identical to y
SuppressDec(xAll, zAll, y)
SuppressDec(xAll, zAll)  # y can be seen in z

# Similar to Y* in paper (but other random values)
SuppressDec(x, z, y)

# Residual standard error forced to be 1
SuppressDec(x, z, y, rmse = 1)

# Five ways of obtaining the same output
SuppressDec(x, z, rmse = 1)  # slower, y must be estimated
SuppressDec(xAll, zAllSupp, y, rmse = 1)
SuppressDec(xAll, zAllSupp, rmse = 1)  # slower, y must be estimated
SuppressDec(xAll, zAll, y, is.na(zAllSupp), rmse = 1)
SuppressDec(xAll, zAll, suppressed = is.na(zAllSupp), rmse = 1)  # y seen in z

# YhatMod4 and YhatMod10 in Table 2 in paper
SuppressDec(xAll, zAllSupp, y, yDeduct = 4 * (y\%/\%4), resScale = 0)
SuppressDec(xAll, zAllSupp, y, yDeduct = 10 * (y\%/\%10), rmse = 0)

# As Table 3 in paper (but other random values)
SuppressDec(xAll, zAllSupp, y, yDeduct = 10 * (y\%/\%10), resScale = 0.1)

# rmse instead of resScale and 5 draws
SuppressDec(xAll, zAllSupp, y, yDeduct = 10 * (y\%/\%10), rmse = 1, nRep = 5)
}
