% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppressDec.R
\name{SuppressDec}
\alias{SuppressDec}
\title{Title}
\usage{
SuppressDec(x, z, y = NULL, digits = 9, nRep = 1, yDeduct = NULL,
  resScale = NULL, rmse = NULL)
}
\arguments{
\item{x}{}

\item{z}{}

\item{y}{}

\item{digits}{}

\item{nRep}{}

\item{yDeduct}{}

\item{resScale}{}

\item{rmse}{}
}
\description{
Title
}
